satrting product carousel administrator view and js, delete file function implemented for programs and about images on micro sites.

+++ app/Http/Controllers/PagesController.php
@@ -574,7 +574,7 @@ class PagesController extends Controller {
             return view('pages.edit.index.edit',$params);
         }
 
-        if($page->page_type_id == 2 || $page->page_type_id == 3) {
+        if($page->page_type_id >= 2) {
 
             if($page->page_type_id == 2) {
                 $page->micro = $page->micro()->first();
@@ -673,10 +673,9 @@ class PagesController extends Controller {
             $page->es_about_1_text = $request->input('es_about_1_text');
             $page->en_about_1_title = $request->input('en_about_1_title');
             $page->en_about_1_text = $request->input('en_about_1_text');
-            // TO TEST: DELETE FILE
-            /*if($request->input('state_about_1_check') == 'removed') {
+            if($request->input('state_about_1_check') == 'removed') {
                 fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','about_1_img'.strchr($page->about_1_img,'.'));
-            }*/
+            }
 
             if($_FILES['about_2_img']['size'] > 0) {
                 $delete = false;
@@ -690,6 +689,9 @@ class PagesController extends Controller {
             $page->es_about_2_text = $request->input('es_about_2_text');
             $page->en_about_2_title = $request->input('en_about_2_title');
             $page->en_about_2_text = $request->input('en_about_2_text');
+            if($request->input('state_about_2_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','about_2_img'.strchr($page->about_2_img,'.'));
+            }
 
             if($_FILES['about_3_img']['size'] > 0) {
                 $delete = false;
@@ -703,6 +705,9 @@ class PagesController extends Controller {
             $page->es_about_3_text = $request->input('es_about_3_text');
             $page->en_about_3_title = $request->input('en_about_3_title');
             $page->en_about_3_text = $request->input('en_about_3_text');
+            if($request->input('state_about_3_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','about_3_img'.strchr($page->about_3_img,'.'));
+            }
 
             //  SEGUNDO BANNER/VIDEO
             if($_FILES['banner_2_img']['size'] > 0) {
@@ -727,6 +732,9 @@ class PagesController extends Controller {
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_1_img',$delete, true,1920,2000,true, false);
                 $page->program_1_img = $_FILES['program_1_img']['name'];
             }
+            if($request->input('state_program_1_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_1_img'.strchr($page->program_1_img,'.'));
+            }
             $page->es_program_1_title = $request->input('es_program_1_title');
             $page->en_program_1_title = $request->input('en_program_1_title');
             $page->es_program_1_text = $request->input('es_program_1_text');
@@ -734,10 +742,6 @@ class PagesController extends Controller {
             if($_FILES['file_program_1']['size'] > 0) {
                 fileUploadController::generalUpload(public_path().'/uploads/pages/'.$page_id.'/','file_program_1',false,true,true);
                 $page->file_program_1 = $_FILES['file_program_1']['name'];
-            }
-            // TO TEST: DELETE FILE
-            if($request->input('state_program_1_check') == 'removed') {
-                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','file_program_1'.strchr($page->file_program_1,'.'));
             }
 
             if($_FILES['program_2_img']['size'] > 0) {
@@ -747,6 +751,9 @@ class PagesController extends Controller {
                 }
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_2_img',$delete, true,1920,2000,true, false);
                 $page->program_2_img = $_FILES['program_2_img']['name'];
+            }
+            if($request->input('state_program_2_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_2_img'.strchr($page->program_2_img,'.'));
             }
             $page->es_program_2_title = $request->input('es_program_2_title');
             $page->en_program_2_title = $request->input('en_program_2_title');
@@ -765,6 +772,9 @@ class PagesController extends Controller {
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_3_img',$delete, true,1920,2000,true, false);
                 $page->program_3_img = $_FILES['program_3_img']['name'];
             }
+            if($request->input('state_program_3_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_3_img'.strchr($page->program_3_img,'.'));
+            }
             $page->es_program_3_title = $request->input('es_program_3_title');
             $page->en_program_3_title = $request->input('en_program_3_title');
             $page->es_program_3_text = $request->input('es_program_3_text');
@@ -781,6 +791,9 @@ class PagesController extends Controller {
                 }
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_4_img',$delete, true,1920,2000,true, false);
                 $page->program_4_img = $_FILES['program_4_img']['name'];
+            }
+            if($request->input('state_program_4_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_4_img'.strchr($page->program_4_img,'.'));
             }
             $page->es_program_4_title = $request->input('es_program_4_title');
             $page->en_program_4_title = $request->input('en_program_4_title');
@@ -799,6 +812,9 @@ class PagesController extends Controller {
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_5_img',$delete, true,1920,2000,true, false);
                 $page->program_5_img = $_FILES['program_5_img']['name'];
             }
+            if($request->input('state_program_5_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_5_img'.strchr($page->program_5_img,'.'));
+            }
             $page->es_program_5_title = $request->input('es_program_5_title');
             $page->en_program_5_title = $request->input('en_program_5_title');
             $page->es_program_5_text = $request->input('es_program_5_text');
@@ -815,6 +831,9 @@ class PagesController extends Controller {
                 }
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_6_img',$delete, true,1920,2000,true, false);
                 $page->program_6_img = $_FILES['program_6_img']['name'];
+            }
+            if($request->input('state_program_6_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_6_img'.strchr($page->program_6_img,'.'));
             }
             $page->es_program_6_title = $request->input('es_program_6_title');
             $page->en_program_6_title = $request->input('en_program_6_title');
@@ -833,6 +852,9 @@ class PagesController extends Controller {
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_7_img',$delete, true,1920,2000,true, false);
                 $page->program_7_img = $_FILES['program_7_img']['name'];
             }
+            if($request->input('state_program_7_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_7_img'.strchr($page->program_7_img,'.'));
+            }
             $page->es_program_7_title = $request->input('es_program_7_title');
             $page->en_program_7_title = $request->input('en_program_7_title');
             $page->es_program_7_text = $request->input('es_program_7_text');
@@ -849,6 +871,9 @@ class PagesController extends Controller {
                 }
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_8_img',$delete, true,1920,2000,true, false);
                 $page->program_8_img = $_FILES['program_8_img']['name'];
+            }
+            if($request->input('state_program_8_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_8_img'.strchr($page->program_8_img,'.'));
             }
             $page->es_program_8_title = $request->input('es_program_8_title');
             $page->en_program_8_title = $request->input('en_program_8_title');
@@ -867,6 +892,9 @@ class PagesController extends Controller {
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_9_img',$delete, true,1920,2000,true, false);
                 $page->program_9_img = $_FILES['program_9_img']['name'];
             }
+            if($request->input('state_program_9_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_9_img'.strchr($page->program_9_img,'.'));
+            }
             $page->es_program_9_title = $request->input('es_program_9_title');
             $page->en_program_9_title = $request->input('en_program_9_title');
             $page->es_program_9_text = $request->input('es_program_9_text');
@@ -883,6 +911,9 @@ class PagesController extends Controller {
                 }
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_10_img',$delete, true,1920,2000,true, false);
                 $page->program_10_img = $_FILES['program_10_img']['name'];
+            }
+            if($request->input('state_program_10_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_10_img'.strchr($page->program_10_img,'.'));
             }
             $page->es_program_10_title = $request->input('es_program_10_title');
             $page->en_program_10_title = $request->input('en_program_10_title');
@@ -901,6 +932,9 @@ class PagesController extends Controller {
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_11_img',$delete, true,1920,2000,true, false);
                 $page->program_11_img = $_FILES['program_11_img']['name'];
             }
+            if($request->input('state_program_11_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_11_img'.strchr($page->program_11_img,'.'));
+            }
             $page->es_program_11_title = $request->input('es_program_11_title');
             $page->en_program_11_title = $request->input('en_program_11_title');
             $page->es_program_11_text = $request->input('es_program_11_text');
@@ -917,6 +951,9 @@ class PagesController extends Controller {
                 }
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_12_img',$delete, true,1920,2000,true, false);
                 $page->program_12_img = $_FILES['program_12_img']['name'];
+            }
+            if($request->input('state_program_12_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_12_img'.strchr($page->program_12_img,'.'));
             }
             $page->es_program_12_title = $request->input('es_program_12_title');
             $page->en_program_12_title = $request->input('en_program_12_title');
@@ -935,6 +972,9 @@ class PagesController extends Controller {
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_13_img',$delete, true,1920,2000,true, false);
                 $page->program_13_img = $_FILES['program_13_img']['name'];
             }
+            if($request->input('state_program_13_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_13_img'.strchr($page->program_13_img,'.'));
+            }
             $page->es_program_13_title = $request->input('es_program_13_title');
             $page->en_program_13_title = $request->input('en_program_13_title');
             $page->es_program_13_text = $request->input('es_program_13_text');
@@ -951,6 +991,9 @@ class PagesController extends Controller {
                 }
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_14_img',$delete, true,1920,2000,true, false);
                 $page->program_14_img = $_FILES['program_14_img']['name'];
+            }
+            if($request->input('state_program_14_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_14_img'.strchr($page->program_14_img,'.'));
             }
             $page->es_program_14_title = $request->input('es_program_14_title');
             $page->en_program_14_title = $request->input('en_program_14_title');
@@ -969,6 +1012,9 @@ class PagesController extends Controller {
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_15_img',$delete, true,1920,2000,true, false);
                 $page->program_15_img = $_FILES['program_15_img']['name'];
             }
+            if($request->input('state_program_15_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_15_img'.strchr($page->program_15_img,'.'));
+            }
             $page->es_program_15_title = $request->input('es_program_15_title');
             $page->en_program_15_title = $request->input('en_program_15_title');
             $page->es_program_15_text = $request->input('es_program_15_text');
@@ -985,6 +1031,9 @@ class PagesController extends Controller {
                 }
                 fileUploadController::imgUpload(public_path().'/uploads/pages/'.$page_id.'/','program_16_img',$delete, true,1920,2000,true, false);
                 $page->program_16_img = $_FILES['program_16_img']['name'];
+            }
+            if($request->input('state_program_16_check') == 'removed') {
+                fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page_id.'/','program_16_img'.strchr($page->program_16_img,'.'));
             }
             $page->es_program_16_title = $request->input('es_program_16_title');
             $page->en_program_16_title = $request->input('en_program_16_title');
@@ -1161,6 +1210,9 @@ class PagesController extends Controller {
         $page->es_about_1_text = $request->input('es_about_1_text');
         $page->en_about_1_title = $request->input('en_about_1_title');
         $page->en_about_1_text = $request->input('en_about_1_text');
+        if($request->input('state_about_1_check') == 'removed') {
+            fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page->page_id.'/','about_1_img'.strchr($page->about_1_img,'.'));
+        }
 
         if($_FILES['about_2_img']['size'] > 0) {
             $delete = false;
@@ -1174,6 +1226,9 @@ class PagesController extends Controller {
         $page->es_about_2_text = $request->input('es_about_2_text');
         $page->en_about_2_title = $request->input('en_about_2_title');
         $page->en_about_2_text = $request->input('en_about_2_text');
+        if($request->input('state_about_2_check') == 'removed') {
+            fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page->page_id.'/','about_2_img'.strchr($page->about_2_img,'.'));
+        }
 
         if($_FILES['about_3_img']['size'] > 0) {
             $delete = false;
@@ -1187,6 +1242,9 @@ class PagesController extends Controller {
         $page->es_about_3_text = $request->input('es_about_3_text');
         $page->en_about_3_title = $request->input('en_about_3_title');
         $page->en_about_3_text = $request->input('en_about_3_text');
+        if($request->input('state_about_3_check') == 'removed') {
+            fileUploadController::deleteFile(public_path().'/uploads/pages/'.$page->page_id.'/','about_3_img'.strchr($page->about_3_img,'.'));
+        }
 
         $page->save();
